Moved list components into right map
